Fixes for backtraces with poorly specified PHP behavior
Sometimes, PHP does not provide a line number and filename in a
debug_backtrace(), including cases where call_user_function() is used.
We must supply fake values, since the function name is still useful to log.

Also, in an Exception's getTrace(), the first frame does not include
a line and filename, but that information can be retrieved from the Exception's
getFile() and getLine() methods.

Lastly, since debug_backtrace() will always include the erroring line,
don't force include it, to avoid an inaccurate and misleading stack trace.