Fix query string guessing in scrubber [Fixes #147]
We are doing a bit of guess work when given a string to scrub to decide
whether we should treat it as a bare query string or just as a string.
This adds some additional tests and changes the way we guess if we have
a query string based on the values of the parsed string being not falsey.