Add error type check before reporting fatal error
Since fatal error handler is connected to script shutdown function
and is executed after every script including ones without fatal errors,
error_get_last() might in some cases return last supressed error that
was not caught by regular error handler. This causes it to be reported
when it should not be according to the configuration.