Fixed crash when getting stack trace from file with unexpected line endings.
For example: if a method in the stack trace was contained in a file with Unix
line endings("\n") on a development Windows machine (where PHP_EOL is "\r\n").
Then the application crashes endlessly, as a new exception is thrown every time
a stack trace is being generated for the last one. Because the program tries
to read line X from an array with only one element (since the split of file
into lines by line breaks failed) which in turn throws another error.

The error was quite severe as it loops on itself endlessly.