Fixed E_NOTICE: Undefined property: Rollbar\Defaults::$defaultGitBranch
When 'allow_exec' is set to false, E_NOTICE errors are generated with every request due to $this->defaultGitBranch variable not being defined because of false condition.

This change implements more robust value caching that is used only when $allowExec parameter is set to true, which is a less suprising behavior.