Add GitHub Action CI workflow
Run PHPUnit on all advertised-as-supported PHP versions, using both
Windows and Linux-based hosts. Since we're not distributing with a
lockfile, also run on the extremes of dependencies: from the lowest
still-compatible version to the highest, most stable. Uses a cache
to keep the composer installs minimized. Suggests notifications for
run output, but commented out until secrets are known.