Support Xdebug 3 in CI
We depend on Xdebug for code coverage in the tests, and optionally
in the code for improved stack traces. One problem with this: we
can't test the code path for "plain traces without Xdebug" without
some configuration flag to ignore Xdebug if it's installed. Setting
that problem aside for now, the immediate problem is that Xdebug 3
was released in late November and now is in release channels. In
this version, the function we use (`xdebug_get_function_stack`)
needs a new INI setting to enable it, which we're not passing in
CI.

In this commit, we attempt to set the INI value using setup-php;
however, this may not work because setup-php adds this flag itself
via the `coverage: xdebug` flag.