doc: add LTS calendar
It's one thing to list a policy; it's entirely another to put that
policy on a calender. In this commit, we concretely document when
various support ends for all current engines. Since it wouldn't be
fair to end support at a time that has since passed, we adopt
tomorrow as the end date: this happens to align nicely with the
natural fixes only date for PHP 5.6, which reached end-of-life on
2018-12-31.