fix: remove blocking branch
When we have a branch a/b, we can't create an offshoot as a/b/c,
because "a/b" is, in the underlying filesystem, a file. It'd be
like trying to do `touch a; mkdir a`, which is clearly a failure.
Thus, the main branch for 1.x development is next/1.x/main and any
feature branches are next/1.x/whatever-you-want. Likewise for 2.x.