fix: branch naming requirements
We want folks to use the next/x.y/ branch structure, so it's clear
where code is meant to land. Also, note what's required for a
contribution to be accepted. Finally, remove the tag instructions
because those don't apply to typical contributors. We'll move this
to another source soon (possibly a composer script, possibly
automation).