test: use more precise assertion
The scheduled test failed because the returned string was not in
the expected array. However, the original assertTrue test didn't
tell us what that returned string was, so we couldn't adjust our
expectations to match. In this commit, we replace with the
arrayContains assertion, which will show us the value received.