fix: test failure because of signature mismatch
When creating mocks based on classes implementing `\Serializable`,
a type signature error was thrown because the mock uses the
signature of the interface for unmocked methods, which differs
from the signature of our classes.

This commit fixes the failures, by aligning our method with the
`\Serializable` interface for the `unserialize` method, but that's
only half the problem: the `serialize` method doesn't always
return a string, as it's required by the interface. A later fix
will address that.