fix: fail to capture local vars in some cases
For performance reasons, PHP 7.4 implemented a new INI setting
that doesn't capture the local variables and set its default value
to off. Consequently, if one requests local var capture AND uses
the engine in its default configuration, then it will silently
fail (no vars captured, when vars exist to be captured).

The fix adopted here is to explicitly tell the engine to capture
variables in the trace when the user requests local var capture.