feat: replace at() matcher
It's deprecated in PHPUnit 10, and we can use a withConsecutive call
to obtain the same behavior with a little less brittleness. While in
there, also remove the unabashed exception swallowing, as it could
lead to false positives (as it did while converting this code).
Instead, we'll specifically mark cases where an exception is expected
just as we would in other test cases.