feat: remove unnecessary $isUncaught param
When the application logging handles an otherwise uncaught error,
exception, or fatality, the logging path enters the "uncaught"
flow. The implementation for the uncaught flow went through the
usual logging mechanism (PSR-3, `log()`) but appended the uncaught
state value to the `log()` signature.

While not a violation of PHP LSP rules, it nonetheless could be
an error when Rollbar is used as the handler, but a user-defined
logger with a different `log()` API is shimmed in-front of the
Rollbar logger.

For broadest compatibility, and keeping with the principal of
least surprise, we remove that extra parameter and instead pass
the state information as part of the exception object itself.

BREAKING CHANGE