feat: be obvious, not obtuse
In the prior version, the function was obviously going to return
`void` (deleted line 39), whereas my change had it returning
whatever the previous handler returns (added line 36). This is
semantically identical, however: [the handler signature is
defined as being `handler ( Throwable $ex ) : void`][1]. So both
are void. But the former is obvious, the latter is obtuse.

I've changed it back to obviously returning void, per the spec.

[1]:https://www.php.net/manual/en/function.set-exception-handler.php