fix: typo in output error message
In the event the API rejected an occurrence during a `log()` call,
and no message was available, a default message containing a
misspelled word - "mesage" - would be emitted in the log. This
change updates the spelling to "message".

This is a breaking change: any tool grepping/searching for the
original misspelling will need to be updated.

BREAKING CHANGE