fix: missed cleanup of synthetic member variable
Release 3.0.0 added a `logUncaught` method to bypass the need to
extend the PSR-defined `log` method. That implementation tagged the
incoming object as "uncaught" and passed it through to the log
method, which then checked for that tagging. However, the pass
through was not properly cleaning up the tagging and there weren't
sufficient test coverage explicitly on the mechanism.

This elaboration on PR 534 adds that coverage.