refac: remove coupling between classes
There was a hidden dependency between `Rollbar\Defaults` and
`Rollbar\Config`: the former depended upon the latter being loaded,
but no way to enforce that was the case. Failing to load the config
prior to the defaults would result in a null included errno, which
would likely prevent logging of any kind.