feat: grandfather current static analysis failures
We want to begin using static analysis to improve the code base,
but we don't want to go back through the 760+ current issues before
doing so. This commit uses the Psalm baseline functionality to
grandfather in the current failures and adds psalm static analysis
to the test cycle. This will catch future issues in the code, while
allowing for gradual remediation of historical issues.