refac: simplify git branch detection
The prior code was wrapped in a try/catch, which wasn't a possible
condition since none of the try {} code was defined as throwing.
Removed the try/catch and called rtrim only when a string was
returned from shell_exec: that was a static analysis failure.

Side note: there's still a problem with this code: shell_exec may
exist, but not be callable -- that needs to be checked for as
well.

@see https://stackoverflow.com/a/21581873/2908724