fix: try using __serialize when obj implements \Serializable
The Serializable interface is now being deprecated in favour of using
the magic methods __serialize() and __unserialize().

Some packages decided to start throwing exceptions on their Serializable
implementations if the interface methods were called.

This commit tries to check if the classes that implement Serializable
are already using __serialize() and if they are, it calls that method,
instead of the interface one serialize().

If this is the case, we also hide the deprecation notice, since it seems
the client code is already preparing to drop \Serializable when the time
comes.

This should fix issue #136 on  rollbar/rollbar-php-laravel